// frontend/src/app/models/country.model.ts
export interface Country {
  id: number;
  name: string;
  capital: string | null;
  population: number | null;
  region: string | null;
  area: number | null;
  flagUrl: string | null;
  cca2: string | null;
  capitalLat: number | null;
  capitalLng: number | null;
  emoji: string | null;
  phoneCode: string | null;
}